+++ test/unit/model/dataset_test.exs
@@ -2340,17 +2340,6 @@ defmodule RDF.DatasetTest do
 
       assert Enum.into(triples, Dataset.new()) == Dataset.new(triples)
     end
-
-    test "with a list of lists" do
-      lists = [
-        [EX.Subject, EX.predicate1(), EX.Object1],
-        [EX.Subject, EX.predicate2(), EX.Object2],
-        [EX.Subject, EX.predicate2(), EX.Object2, EX.Graph]
-      ]
-
-      assert Enum.into(lists, Dataset.new()) ==
-               Dataset.new(Enum.map(lists, &List.to_tuple/1))
-    end
   end
 
   describe "Access behaviour" do